refactor: group `Popup` related componentes in dedicated directory
